+++ scripts/adblock/enable.sh
@@ -7,6 +7,8 @@ source_script=$cwd/$update_script
 
 cron_script=/etc/cron.weekly/$update_script
 
-ln -s $source_script $cron_script
+if [ ! -f $cron_script ]; then
+	ln -s $source_script $cron_script
+fi
 
 $source_script
